+++ guix-gui/src/main.rs
@@ -5,6 +5,7 @@ mod app_metadata;
 mod carrier;
 mod operation_subscription;
 mod progress_summary;
+mod recommended;
 mod settings;
 mod styles;
 mod terminal_buffer;